+++ contracts/bridge/contracts/interfaces/IOutboundLane.sol
@@ -2,5 +2,5 @@
 pragma solidity >=0.6.0 <0.7.0;
 
 interface IOutboundLane {
-    function send_message(address targetContract, bytes calldata encoded) external payable returns (uint256);
+    function send_message(address targetContract, bytes calldata encoded) external payable returns (uint64);
 }
